% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStability.R
\name{checkStability}
\alias{checkStability}
\title{Function to make a Stabilisation Graph from matrices of total costs and QALYs}
\usage{
checkStability(
  total_costs = example_TC,
  total_qalys = example_TQ,
  strategies = NULL,
  ci_lower = 0.025,
  ci_upper = 0.975,
  BS_samples = 500,
  withinShiny = F,
  lambda = 20000,
  plot_CnQ = T
)
}
\arguments{
\item{total_costs}{A matrix of costs for PSA runs.}

\item{total_qalys}{A matrix of QALYs for PSA runs.}

\item{strategies}{A vector of strategy names to include in analysis, must be same as the column names.}

\item{ci_lower}{Lower bound of the confidence interval.}

\item{ci_upper}{Upper of the confidence interval.}

\item{BS_samples}{number of samples to use in the bootstrap - default 1000}

\item{withinShiny}{TRUE if used within Shiny app.}

\item{lambda}{WTP used in Net benefit lot.}

\item{plot_CnQ}{if T then plots costs and qalys alongside net benefit, else plots only net-benefit.}
}
\value{
a PSA stabilisation plot for costs and qalys.
}
\description{
Function to make a Stabilisation Graph from matrices of total costs and QALYs
}
\examples{
checkStability(total_costs = example_TC, total_qalys = example_TQ)
}
