\name{darkpeak-package}
\alias{darkpeak-package}
\alias{darkpeak}
\docType{package}
\title{
\packageTitle{darkpeak}
}
\description{
\packageDescription{darkpeak}
}
\details{
For more details see \code{\link[Dark Peak Analytics]{https://darkpeakanalytics.com}}
}
\author{
\packageAuthor{darkpeak}

Maintainer: \packageMaintainer{darkpeak}
}
\references{
Smith, R. and Schneider, P., 2020. Making health economic models Shiny: A tutorial. Wellcome Open Research, 5(69), p.69.
}
\keyword{ package }
\seealso{
~~ \code{\link[Rcpp]{https://cran.r-project.org/web/packages/Rcpp/index.html}} ~~
}
\examples{
See individual functions for examples or contact Rob.
}
