% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/darkpeak_cols.R
\name{darkpeak_cols}
\alias{darkpeak_cols}
\title{Function to extract darkpeak colors as hex codes}
\usage{
darkpeak_cols(...)
}
\arguments{
\item{...}{Character names of darkpeak_colors}
}
\description{
Function to extract darkpeak colors as hex codes
}
