% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCEPlane.R
\name{makeCEPlane}
\alias{makeCEPlane}
\title{Function to make a Cost-effectiveness Plane plot from matrices of total costs and QALYs}
\usage{
makeCEPlane(total_costs = total_costs,total_qalys = total_qalys,comparitor = "treat 1",treatment = "notreat",thresh = 20000,show_ellipse = T)
}
\arguments{
\item{total_costs}{A matrix of costs for PSA runs.}

\item{total_qalys}{A matrix of QALYs for PSA runs.}

\item{treatment}{A vector of treatment names (must match matrix column names).}
}
\value{
a CE-Plane plot produced in ggplot.
}
\description{
Function to make a Cost-effectiveness Plane plot from matrices of total costs and QALYs
}
\examples{
makeCEPlane(total_costs = example_TC,total_qalys = example_TQ,comparitor = "treat 1",treatment = "notreat",thresh = 20000,show_ellipse = T)
}
