\name{darkpeak-package}
\alias{darkpeak-package}
\alias{darkpeak}
\docType{package}
\title{
\packageTitle{darkpeak}
}
\description{
\packageDescription{darkpeak}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{darkpeak}
\packageIndices{darkpeak}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{darkpeak}

Maintainer: \packageMaintainer{darkpeak}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
