% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCEAC.R
\name{makeCEAC}
\alias{makeCEAC}
\title{Function to make a CEAC plot from matrices of total costs and QALYs}
\usage{
makeCEAC(
  total_costs,
  total_qalys,
  treatment = c("treat 1","notreat")
)
}
\arguments{
\item{total_costs}{A matrix of costs for PSA runs.}

\item{total_qalys}{A matrix of QALYs for PSA runs.}

\item{treatment}{A vector of treatment names (must match matrix column names).}
}
\value{
a CEAC plot produced in ggplot.
}
\description{
Function to make a CEAC plot from matrices of total costs and QALYs
}
\examples{
makeCEAC(total_costs = example_TC,total_qalys = example_TQ,treatment = c("treat 1","notreat"))
}
