% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCEPlane.R
\name{makeCEPlane}
\alias{makeCEPlane}
\title{Function to make a Cost-effectiveness Plane plot from matrices of total costs and QALYs}
\usage{
makeCEPlane(
  total_costs = NULL,
  total_qalys = NULL,
  comparitor = "Coumarin (INR 2-3)",
  treatment = c("Apixaban (5mg bd)", "Dabigatran (150mg bd)", "Edoxaban (60mg od)",
    "Rivaroxaban (20mg od)", "No treatment"),
  thresh = NULL,
  show_ellipse = F
)
}
\arguments{
\item{total_costs}{A matrix of costs for PSA runs.}

\item{total_qalys}{A matrix of QALYs for PSA runs.}

\item{treatment}{A vector of treatment names (must match matrix column names).}
}
\value{
a CE-Plane plot produced in ggplot.
}
\description{
Function to make a Cost-effectiveness Plane plot from matrices of total costs and QALYs
}
\examples{
makeCEPlane(total_costs = m_costs,
total_qalys = m_qalys,
treatment = c("Apixaban (5mg bd)","Dabigatran (150mg bd)","Coumarin (INR 2-3)", "Edoxaban (60mg od)","Rivaroxaban (20mg od)", "No treatment"),
comparitor = "Coumarin (INR 2-3)")
}
